+++ CDC/Models/SR1UnitModel.cs
@@ -241,10 +241,10 @@ namespace CDC
 			int v3 = reader.ReadUInt16();
 
 			// unsigned char attr;
-			byte attr = reader.ReadByte();
+			ushort attr = reader.ReadByte();
 			if (options.IgnoreBackfacingFlagForTerrain)
 			{
-				attr &= (byte)~PolygonFlags.Backfacing;
+				attr &= (ushort)~PolygonFlags.Backfacing;
 			}
 
 			// char sortPush;
@@ -328,17 +328,17 @@ namespace CDC
 				return;
 			}
 
-			if ((material.polygonFlags & (byte)PolygonFlags.Hidden0) != 0)
+			if ((material.polygonFlags & (ushort)PolygonFlags.Hidden0) != 0)
 			{
 				material.visible = false;
 			}
 
-			if ((material.polygonFlags & (byte)PolygonFlags.Emissive) != 0)
+			if ((material.polygonFlags & (ushort)PolygonFlags.Emissive) != 0)
 			{
 				material.isEmissive = true;
 			}
 
-			if ((material.polygonFlags & (byte)PolygonFlags.Translucent) != 0)
+			if ((material.polygonFlags & (ushort)PolygonFlags.Translucent) != 0)
 			{
 				material.isTranslucent = true;
 			}
